se cargan las coordenadas al marcar el mapa, se modifico tipo de dato de coordenadas en el back

+++ back/sensores/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
+from sqlalchemy import Boolean, Column, ForeignKey, Integer, String,Float
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 from enum import auto, StrEnum
 from datetime import datetime, UTC
@@ -12,6 +12,5 @@ class Sensor(BaseModel):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, index=True)
     identificador: Mapped[str] = mapped_column(String, index=True)
     porcentajeBateria: Mapped[int] = mapped_column(Integer,index = True)
-    latitud: Mapped[int] = mapped_column(Integer,index = True, nullable=True)
-    longitud: Mapped[int] = mapped_column(Integer,index = True, nullable=True)
-
+    latitud: Mapped[float] = mapped_column(Float, index=True, nullable=True)  
+    longitud: Mapped[float] = mapped_column(Float, index=True, nullable=True)
